Use OSGi annotation classes
Rather than using Felix annotations, this makes use of the official
OSGi annotations. As such, `@Reference` fields are now accessible
with `setX` and `unsetX` methods for use with OSGi declarative services.
And `@Service` and `@Property` annotations have been merged with the
`@Component` declaration.

The Felix maven-scr-plugin is updated to the latest version.

The official JAX-RS dependency is also updated.